requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
